feat: add handler folder to group handle inputs
