+++ cli/src/commands/tui/ui/logs.rs
@@ -1,3 +1,5 @@
+use super::util::get_color;
+use crate::commands::tui::app::{ActiveBlock, App, State, MAX_LOG_ENTRIES_LENGTH};
 use ratatui::{
     prelude::{Alignment, Backend, Constraint, Direction, Layout, Margin, Rect},
     style::{Color, Style},
@@ -8,14 +10,6 @@ use ratatui::{
     Frame,
 };
 use wukong_sdk::services::gcloud::google::logging::r#type::LogSeverity;
-
-use crate::commands::tui::{
-    action::Action,
-    app::{ActiveBlock, App, AppReturn, State, MAX_LOG_ENTRIES_LENGTH},
-    events::{key::Key, network::NetworkEvent},
-};
-
-use super::util::get_color;
 
 pub struct LogsWidget;
 
@@ -53,80 +47,6 @@ impl LogsWidget {
 
         render_log_entries(frame, logs_area, &mut app.state);
         render_scrollbar(frame, logs_area, &mut app.state.logs_vertical_scroll_state);
-    }
-
-    pub async fn handle_input(key: Key, app: &mut App) -> AppReturn {
-        match Action::from_key(key) {
-            Some(Action::ToggleLogsTailing) => {
-                app.state.logs_tailing = !app.state.logs_tailing;
-            }
-            Some(Action::ShowErrorAndAbove) => {
-                app.dispatch(NetworkEvent::GetGCloudLogs).await;
-
-                app.state.is_fetching_log_entries = true;
-                app.state.start_polling_log_entries = false;
-
-                app.state.log_entries = vec![];
-                app.state.log_entries_length = 0;
-                // Need to reset scroll, or else it will be out of bound
-
-                // Add if not already in the list
-                // or else remove it
-                app.state.logs_severity = match app.state.logs_severity {
-                    Some(LogSeverity::Error) => None,
-                    _ => Some(LogSeverity::Error),
-                };
-            }
-            _ => match key {
-                Key::Esc | Key::Char('q') => {
-                    app.pop_navigation_stack();
-                }
-                // Key::Enter => handle_enter_key(app).await,
-                Key::Up | Key::Char('k') => {
-                    app.state.logs_vertical_scroll =
-                        app.state.logs_vertical_scroll.saturating_sub(5);
-                    app.state.logs_vertical_scroll_state = app
-                        .state
-                        .logs_vertical_scroll_state
-                        .position(app.state.logs_vertical_scroll);
-
-                    app.state.logs_enable_auto_scroll_to_bottom = false;
-                }
-                Key::Down | Key::Char('j') => {
-                    app.state.logs_vertical_scroll =
-                        app.state.logs_vertical_scroll.saturating_add(5);
-                    app.state.logs_vertical_scroll_state = app
-                        .state
-                        .logs_vertical_scroll_state
-                        .position(app.state.logs_vertical_scroll);
-
-                    app.state.logs_enable_auto_scroll_to_bottom = false;
-                }
-                Key::Left | Key::Char('h') => {
-                    app.state.logs_horizontal_scroll =
-                        app.state.logs_horizontal_scroll.saturating_sub(5);
-                    app.state.logs_horizontal_scroll_state = app
-                        .state
-                        .logs_horizontal_scroll_state
-                        .position(app.state.logs_horizontal_scroll);
-
-                    app.state.logs_enable_auto_scroll_to_bottom = false;
-                }
-                Key::Right | Key::Char('l') => {
-                    app.state.logs_horizontal_scroll =
-                        app.state.logs_horizontal_scroll.saturating_add(5);
-                    app.state.logs_horizontal_scroll_state = app
-                        .state
-                        .logs_horizontal_scroll_state
-                        .position(app.state.logs_horizontal_scroll);
-
-                    app.state.logs_enable_auto_scroll_to_bottom = false;
-                }
-                _ => {}
-            },
-        };
-
-        AppReturn::Continue
     }
 }
