+++ cli/src/commands/tui/ui/namespace_selection.rs
@@ -1,3 +1,5 @@
+use super::{centered_rect, create_loading_widget};
+use crate::commands::tui::app::App;
 use ratatui::{
     prelude::{Alignment, Backend},
     style::{Color, Modifier, Style},
@@ -5,13 +7,6 @@ use ratatui::{
     widgets::{Block, Borders, Clear, List, ListItem},
     Frame,
 };
-
-use crate::commands::tui::{
-    app::{ActiveBlock, App, AppReturn},
-    events::{key::Key, network::NetworkEvent},
-};
-
-use super::{centered_rect, create_loading_widget};
 
 pub struct NamespaceSelectionWidget;
 
@@ -54,42 +49,4 @@ impl NamespaceSelectionWidget {
             frame.render_stateful_widget(items, area, &mut app.namespace_selections.state);
         }
     }
-
-    pub async fn handle_input(key: Key, app: &mut App) -> AppReturn {
-        match key {
-            Key::Up => app.namespace_selections.previous(),
-            Key::Down => app.namespace_selections.next(),
-            Key::Esc | Key::Char('q') => app.push_navigation_stack(ActiveBlock::Empty),
-            Key::Enter => {
-                let selected = app
-                    .namespace_selections
-                    .items
-                    .get(app.namespace_selections.state.selected().unwrap())
-                    .unwrap();
-
-                if let Some(current_namespace) = &app.state.current_namespace {
-                    if current_namespace != selected {
-                        fetch_and_reset_polling(app, selected.to_string()).await;
-                    }
-                } else {
-                    fetch_and_reset_polling(app, selected.to_string()).await;
-                }
-
-                app.push_navigation_stack(ActiveBlock::Empty)
-            }
-            _ => {}
-        };
-
-        AppReturn::Continue
-    }
-}
-
-async fn fetch_and_reset_polling(app: &mut App, selected_version: String) {
-    app.state.current_namespace = Some(selected_version);
-    app.dispatch(NetworkEvent::GetBuilds).await;
-    app.dispatch(NetworkEvent::GetGCloudLogs).await;
-
-    app.state.is_fetching_log_entries = true;
-    app.state.start_polling_log_entries = false;
-    app.state.has_log_errors = false;
 }
