+++ cli/src/commands/tui/ui/version_selection.rs
@@ -1,3 +1,5 @@
+use super::{centered_rect, create_loading_widget};
+use crate::commands::tui::app::App;
 use ratatui::{
     prelude::{Alignment, Backend},
     style::{Color, Modifier, Style},
@@ -5,13 +7,6 @@ use ratatui::{
     widgets::{Block, Borders, Clear, List, ListItem},
     Frame,
 };
-
-use crate::commands::tui::{
-    app::{ActiveBlock, App, AppReturn},
-    events::{key::Key, network::NetworkEvent},
-};
-
-use super::{centered_rect, create_loading_widget};
 
 pub struct VersionSelectionWidget;
 
@@ -53,45 +48,4 @@ impl VersionSelectionWidget {
             frame.render_stateful_widget(items, area, &mut app.version_selections.state);
         }
     }
-
-    pub async fn handle_input(key: Key, app: &mut App) -> AppReturn {
-        match key {
-            Key::Esc | Key::Char('q') => app.push_navigation_stack(ActiveBlock::Empty),
-            Key::Enter => handle_enter_key(app).await,
-            Key::Up => app.version_selections.previous(),
-            Key::Down => app.version_selections.next(),
-            _ => {}
-        };
-
-        AppReturn::Continue
-    }
-}
-
-async fn handle_enter_key(app: &mut App) {
-    let selected_version_index = match app.version_selections.state.selected() {
-        Some(index) => index,
-        None => return, // No selected version, nothing to do
-    };
-
-    let selected_version = app.version_selections.items[selected_version_index].clone();
-
-    if let Some(current_namespace) = &app.state.current_namespace {
-        if selected_version == *current_namespace {
-            app.push_navigation_stack(ActiveBlock::Empty);
-            return;
-        }
-    }
-
-    fetch_and_reset_polling(app, selected_version).await;
-    app.push_navigation_stack(ActiveBlock::Empty);
-}
-
-async fn fetch_and_reset_polling(app: &mut App, selected_version: String) {
-    app.state.current_version = Some(selected_version);
-
-    app.state.is_fetching_log_entries = true;
-    app.state.start_polling_log_entries = false;
-    app.state.has_log_errors = false;
-
-    app.dispatch(NetworkEvent::GetBuilds).await;
 }
